PanelData: move preProcessPanelData to @grafana/data
