+++ public/app/features/alerting/unified/state/AlertingQueryRunner.ts
@@ -11,13 +11,13 @@ import {
   rangeUtil,
   TimeRange,
   withLoadingIndicator,
+  preProcessPanelData,
 } from '@grafana/data';
 import { FetchResponse, getDataSourceSrv, toDataQueryError } from '@grafana/runtime';
 import { BackendSrv, getBackendSrv } from 'app/core/services/backend_srv';
 import { isExpressionQuery } from 'app/features/expressions/guards';
 import { cancelNetworkRequestsOnUnsubscribe } from 'app/features/query/state/processing/canceler';
 import { setStructureRevision } from 'app/features/query/state/processing/revision';
-import { preProcessPanelData } from 'app/features/query/state/runRequest';
 import { AlertQuery } from 'app/types/unified-alerting-dto';
 
 import { getTimeRangeForExpression } from '../utils/timeRange';
