+++ public/app/features/query/state/runRequest.ts
@@ -7,7 +7,6 @@ import { catchError, map, mapTo, share, takeUntil, tap } from 'rxjs/operators';
 // Types
 import {
   CoreApp,
-  DataFrame,
   DataQueryError,
   DataQueryRequest,
   DataQueryResponse,
@@ -15,11 +14,9 @@ import {
   DataSourceApi,
   DataTopic,
   dateMath,
-  guessFieldTypes,
   LoadingState,
   PanelData,
   TimeRange,
-  toDataFrame,
 } from '@grafana/data';
 import { toDataQueryError } from '@grafana/runtime';
 import { isExpressionReference } from '@grafana/runtime/src/utils/DataSourceWithBackend';
@@ -200,60 +197,4 @@ export function callQueryMethod(
   // Otherwise it is a standard datasource request
   const returnVal = queryFunction ? queryFunction(request) : datasource.query(request);
   return from(returnVal);
-}
-
-function getProcessedDataFrame(data: DataQueryResponseData): DataFrame {
-  const dataFrame = guessFieldTypes(toDataFrame(data));
-
-  if (dataFrame.fields && dataFrame.fields.length) {
-    // clear out the cached info
-    for (const field of dataFrame.fields) {
-      field.state = null;
-    }
-  }
-
-  return dataFrame;
-}
-
-/**
- * All panels will be passed tables that have our best guess at column type set
- *
- * This is also used by PanelChrome for snapshot support
- */
-export function getProcessedDataFrames(results?: DataQueryResponseData[]): DataFrame[] {
-  if (!results || !isArray(results)) {
-    return [];
-  }
-
-  return results.map((data) => getProcessedDataFrame(data));
-}
-
-export function preProcessPanelData(data: PanelData, lastResult?: PanelData): PanelData {
-  const { series, annotations } = data;
-
-  //  for loading states with no data, use last result
-  if (data.state === LoadingState.Loading && series.length === 0) {
-    if (!lastResult) {
-      lastResult = data;
-    }
-
-    return {
-      ...lastResult,
-      state: LoadingState.Loading,
-      request: data.request,
-    };
-  }
-
-  // Make sure the data frames are properly formatted
-  const STARTTIME = performance.now();
-  const processedDataFrames = series.map((data) => getProcessedDataFrame(data));
-  const annotationsProcessed = getProcessedDataFrames(annotations);
-  const STOPTIME = performance.now();
-
-  return {
-    ...data,
-    series: processedDataFrames,
-    annotations: annotationsProcessed,
-    timings: { dataProcessingTime: STOPTIME - STARTTIME },
-  };
 }
